Improve handling of dead keys
When you press the dead key "^" followed by space on a German keyboard,
the applicaton receives the following two KeyDown events:
1. event.key = Key_Circumflex, event.character = ""
2. event.key = Key_Space, event.character = "^"

Under Gtk this should work for all dead keys by using GtkIMContext.
Under Windows it is implemented only for the scan code of the dead key
"^" on a German keyboard.